Validate motor model only in motor_model.py. Combine V and fric model graphs. Remove import *.
